services/voice_gateway/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
PyJWT==2.8.0

# Voice services - STT
assemblyai
# ElevenLabs uses REST API directly (via requests)

# Messaging
aio-pika==9.3.0

# Database
asyncpg==0.29.0

# Circuit breaker
pybreaker==1.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Audio processing - audioop is deprecated in Python 3.11, removed in 3.13
audioop-lts==0.2.1;python_version>="3.13"

# Audio format conversion (MP3 to PCM)
pydub==0.25.1
